requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
tensorflow>=2.13
matplotlib>=3.7
joblib>=1.3
openpyxl>=3.1
xlrd>=2.0
scipy>=1.10
pillow>=10.0
